Clarity over Midnight scheme ancestry
Root cause of confusion is inconsistent versioning the addition of
ALT SARSCoV_1200_28_LEFT_27837T [0] to accomodate BA.1, whose inclusion
led to a version bump by ONT but not by IDT and thus versions getting
out of step. The first kits from each vendor to include this primer
are midnight-ont-v2 and midnight-bccdc-v4, of which midnight-idt-v1
is an alias.
[0] https://twitter.com/freed_nikki/status/1464477522448433156